api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-JWT-Extended
pydantic
python-dotenv
Werkzeug
flask-restx
